# api/swagger.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: 'Association'
  description: 'This is voluntary association'
  contact:
    email: "leosmile38@gmail.com"
  license:
    name: "Apache-2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "youthspeakshawaii.herokuapp.com"
...
basePath: "/v2"
schemes:
  - "https"
  - "http"
...
paths:
...
  /events/{id}:
    get:
      tags:
        - "event"
      summary: "Find event with the event id"
      operationId: "eventIdGET"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "id of the event you are searching"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "operation was successful"
          schema:
            $ref: '#/definitions/Event'
      x-swagger-router-controller: "Event"
  /events/{id}/contact:
    get:
      tags:
        - "event"
      summary: "Get the contact of an event"
      operationId: "eventsIdContactGET"
      parameters:
        - name: "id"
          in: "path"
          description: "id of the event you want the contact of"
          required: true
          type: "integer"
          format: "int64"
      responses:
        '200':
          description: "successful operation"
          schema:
            $ref: '#/definitions/Contact'
      x-swagger-router-controller: "Contact"
  /events/{id}/activity:
    get:
      tags:
        - "book"
      summary: "Get the activity of an event"
      operationId: "eventsIdActivityGET"
      parameters:
        - name: "id"
          in: "path"
          description: "id of the event you want the activity of"
          required: true
          type: "integer"
          format: "int64"
      responses:
        '200':
          description: "successful operation"
          schema:
            $ref: '#/definitions/Activity'

      x-swagger-router-controller: "Event"
      
  # /events/{id}/similar:
  #   get:
  #     tags:
  #       - "book"
  #     summary: "Get similar books to the chosen one"
  #     operationId: "booksIdSimilarGET"
  #     parameters:
  #       - name: "id"
  #         in: "path"
  #         description: "id of the book you want the similars of"
  #         required: true
  #         type: "integer"
  #         format: "int64"
  #     responses:
  #       '200':
  #         description: "successful operation"
  #         schema:
  #           $ref: '#/definitions/Book'
  #     x-swagger-router-controller: "Book"
      
  # /books/themes:
  #   get:
  #     tags:
  #       - "book"
  #     summary: "Get all the themes that are present"
  #     operationId: "booksThemesGET"
  #     produces:
  #       - "application/json"
  #     responses:
  #       '200':
  #         description: "All themes"
  #     x-swagger-router-controller: "Book"

  # /books/themes/{theme}:
  #   get:
  #     tags:
  #       - "book"
  #     summary: "Get books with a given theme"
  #     operationId: "booksByThemeGET"
  #     parameters:
  #       - name: "theme"
  #         in: "path"
  #         description: "theme of the book you want"
  #         required: true
  #         type: "string"
  #     produces:
  #       - "application/json"
  #     responses:
  #       '200':
  #         description: "successful operation"
  #     x-swagger-router-controller: "Book"

  # /books/genres:
  #   get:
  #     tags:
  #       - "book"
  #     summary: "Get all the genres that are present"
  #     operationId: "booksGenresGET"
  #     produces:
  #       - "application/json"
  #     responses:
  #       '200':
  #         description: "successful operation"
  #     x-swagger-router-controller: "Book"

  # /books/genres/{genre}:
  #   get:
  #     tags:
  #       - "book"
  #     summary: "Get books with a given genre"
  #     operationId: "booksByGenreGET"
  #     parameters:
  #       - name: "genre"
  #         in: "path"
  #         description: "genre of the book you want"
  #         required: true
  #         type: "string"
  #     produces:
  #       - "application/json"
  #     responses:
  #       '200':
  #         description: "successful operation"
  #         schema:
  #           type: "array"
  #           items:
  #             $ref: "#/definitions/Book"
  #     x-swagger-router-controller: "Book"
...
definitions:
  Event:
    type: "object"
    required:
      - "id"
      - "name"
      - "activityId"
      - "date"
      - "location"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      activityId:
        type: "integer"
        format: "int64"
      date:
        type: "string"
        format: "date"
      location:
        type: "string"
      descritpion:
        type: "string"
    xml:
      name: "Event"
    title: "Event"
    description: "An event related to an activity"
    example:
      id: 0
      name: "How to create a poem"
      date: "2020-07-21"
      location: "Honolulu"
      description: "Expressing your thoughts"


  # Book:
  #   type: "object"
  #   required:
  #     - "id"
  #     - "title"
  #     - "date"
  #   properties:
  #     id:
  #       type: "integer"
  #       format: "int64"
  #     title:
  #       type: "string"
  #     price:
  #       type: "number"
  #       format: "double"
  #     isbn:
  #       type: "string"
  #     genre:
  #       type: "string"
  #     quantity:
  #       type: "integer"
  #       format: "int32"
  #     description:
  #       type: "string"
  #     publisher:
  #       type: "string"
  #     language:
  #       type: "string"
  #     release:
  #       type: "string"
  #       format: "date"
  #     field:
  #       type: "string"
  #     theme:
  #       type: "string"

  #   xml:
  #     name: "Book"
  #   title: "Book"
  #   description: "A book for sale in the store"
  #   example:
  #     id: 0
  #     title: "Harry Potter 1"
  #     price: 10
  #     isbn: "9780747532743"
  #     genre: "fantasy"
  #     quantity: 3
  #     publisher: "Bloomsbury Publishing"
  #     language: "english"
  #     release: "1997-06-26"
  #     field: "Wizardry"
  #     theme: "Harry Potter Exhibition"
  Activity:
    type: "object"
    required:
      - "id"
      - "name"
      - "eventId"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      eventId:
        type: "integer"
        format: "int64"
      descritpion:
        type: "string"
    xml:
      name: "Activity"
    title: "Activity"
    description: "An activity related to an event"
    example:
      id: 0
      name: "Poetry 101"
      description: "Helping children to create short poems"
  Contact:
    type: "object"
    required:
      - "id"
      - "name"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      surname:
        type: "string"
      info:
        type: "string"
    xml:
      name: "Contact"
    description: "Volunteer"
    example:
      id: 5
      name: "Jason"
      surname: "Momoa"
      info: "Actor and activist"

  # User:
  #   type: "object"
  #   required:
  #     - "email"
  #     - "password"
  #   properties:
  #     email:
  #       type: "string"
  #     password:
  #       type: "string"
  #   xml:
  #     name: "User"
  #   example:
  #     email: "harry.potter@gmail.com"
  #     password: "wingardiumleviosa"